feat: HeaderService에서 직접 API 요청하는 것이 아닌 PostId에서 Prop으로 전달 / PostId에서 배경색뿐만 아니라 배경이미지도 적용
